moved interfaces in params.ts to the bottom of database.ts and edited exports in index.ts
