+++ server/src/controllers/getCourse.ts
@@ -2,7 +2,7 @@ import { Response, Request } from 'express'
 import * as util from 'util'
 
 import Database from '../database'
-import { dbReadParams } from '../../interfaces/params'
+import { dbReadParams } from '../database'
 import { GetCourseParams } from '../../interfaces/GetCourseParams'
 
 /**
